Consolidate actor-claim benchmarks into descriptive classes (PR #3571 feedback)
Address review feedback (Toidi357): combine the two actor-claim benchmark files
and give the classes descriptive, parallel names.

- Merge ActorClaimBenchmarks.cs + ActorClaimReadBenchmarks.cs into one file with
  a shared ActorClaimBenchmarkData helper (Descriptor, BuildSubjectWithActorChain),
  removing the duplicated builders.
- Rename classes: ActorClaimSerializationBenchmarks (write path) and
  ActorClaimDeserializationBenchmarks (read path). Benchmarks are unchanged.
- Update the ActorClaim.md benchmark reference.
