+++ benchmark/Microsoft.IdentityModel.Benchmarks/ActorClaimBenchmarks.cs
@@ -1,46 +1,30 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System.Collections.Generic;
 using System.Security.Claims;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Microsoft.IdentityModel.JsonWebTokens;
 using Microsoft.IdentityModel.Tokens;
 
 namespace Microsoft.IdentityModel.Benchmarks;
 
-// dotnet run -c release -f net8.0 --filter Microsoft.IdentityModel.Benchmarks.ActorClaimBenchmarks*
+// Actor ("act") claim benchmarks. The serialization (write-path) and deserialization (read-path)
+// benchmarks share the token/subject builders in ActorClaimBenchmarkData below.
 //
-// Serialization (write-path) benchmarks for actor ("act") claim support: CreateToken only.
+//   dotnet run -c release -f net8.0 --filter *ActorClaim*Benchmarks*
 //
-// "MaxDepthN_NestedM" means JsonWebTokenHandler.MaxActorChainLength == N with an actor chain M
-// levels deep. When M > N the chain overflows and, on write, the remainder is serialized as a
-// JSON-text string rather than nested "act" objects. MaxActorChainLength is a process-wide static,
-// so every benchmark sets it explicitly (the assignment is ~1ns and negligible against the
-// microsecond-scale token operations).
-public class ActorClaimBenchmarks
+// JsonWebTokenHandler.MaxActorChainLength is a process-wide static that bounds how deep the actor
+// chain is (de)serialized; every benchmark sets it explicitly (the assignment is ~1ns and negligible
+// against the microsecond-scale token operations). When the chain is deeper than the limit it
+// "degrades": on write the overflow is serialized as a JSON-text string rather than nested "act"
+// objects, and on read that string is kept as a claim rather than expanded.
+
+// Shared token/subject builders for the actor-claim benchmarks.
+internal static class ActorClaimBenchmarkData
 {
-    private JsonWebTokenHandler _jsonWebTokenHandler;
-    private SecurityTokenDescriptor _descNoActor;
-    private SecurityTokenDescriptor _descNested1;
-    private SecurityTokenDescriptor _descNested2;
-    private SecurityTokenDescriptor _descNested4;
-    private SecurityTokenDescriptor _descNested5;
-
-    [GlobalSetup]
-    public void Setup()
-    {
-        _jsonWebTokenHandler = new JsonWebTokenHandler();
-        _descNoActor = Descriptor(BuildSubjectWithActorChain(0));
-        _descNested1 = Descriptor(BuildSubjectWithActorChain(1));
-        _descNested2 = Descriptor(BuildSubjectWithActorChain(2));
-        _descNested4 = Descriptor(BuildSubjectWithActorChain(4));
-        _descNested5 = Descriptor(BuildSubjectWithActorChain(5));
-    }
-
-    [GlobalCleanup]
-    public void Cleanup() => JsonWebTokenHandler.MaxActorChainLength = 1;
-
-    private static SecurityTokenDescriptor Descriptor(ClaimsIdentity subject) => new SecurityTokenDescriptor
+    internal static SecurityTokenDescriptor Descriptor(ClaimsIdentity subject) => new SecurityTokenDescriptor
     {
         Subject = subject,
         Issuer = BenchmarkUtils.Issuer,
@@ -50,7 +34,7 @@ public class ActorClaimBenchmarks
 
     // Builds a subject whose ClaimsIdentity.Actor chain is <paramref name="actorLevels"/> deep
     // (0 = no actor).
-    private static ClaimsIdentity BuildSubjectWithActorChain(int actorLevels)
+    internal static ClaimsIdentity BuildSubjectWithActorChain(int actorLevels)
     {
         var subject = new CaseSensitiveClaimsIdentity(new[] { new Claim("sub", "user-1"), new Claim("name", "User One") });
         var current = subject;
@@ -63,6 +47,35 @@ public class ActorClaimBenchmarks
 
         return subject;
     }
+}
+
+// Serialization (write-path) benchmarks for actor ("act") claim support: CreateToken only.
+//
+// "MaxDepthN_NestedM" means MaxActorChainLength == N with an actor chain M levels deep. When M > N the
+// chain overflows and, on write, the remainder is serialized as a JSON-text string.
+[MemoryDiagnoser]
+public class ActorClaimSerializationBenchmarks
+{
+    private JsonWebTokenHandler _jsonWebTokenHandler;
+    private SecurityTokenDescriptor _descNoActor;
+    private SecurityTokenDescriptor _descNested1;
+    private SecurityTokenDescriptor _descNested2;
+    private SecurityTokenDescriptor _descNested4;
+    private SecurityTokenDescriptor _descNested5;
+
+    [GlobalSetup]
+    public void Setup()
+    {
+        _jsonWebTokenHandler = new JsonWebTokenHandler();
+        _descNoActor = ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(0));
+        _descNested1 = ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(1));
+        _descNested2 = ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(2));
+        _descNested4 = ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(4));
+        _descNested5 = ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(5));
+    }
+
+    [GlobalCleanup]
+    public void Cleanup() => JsonWebTokenHandler.MaxActorChainLength = 1;
 
     [Benchmark(Baseline = true)]
     public string NoActor()
@@ -97,5 +110,94 @@ public class ActorClaimBenchmarks
     {
         JsonWebTokenHandler.MaxActorChainLength = 5;
         return _jsonWebTokenHandler.CreateToken(_descNested4);
+    }
+}
+
+// Deserialization (read-path) benchmarks for actor ("act") claim support: ValidateTokenAsync, then
+// materialize ClaimsIdentity (which is built lazily) so the actor-expansion cost is measured.
+//
+// Covers: no actor, a single "act", a degraded "act" chain (chain deeper than MaxActorChainLength, so
+// the overflow was written as a JSON-text string and is kept as a claim on read), and the legacy
+// "actort" (unsigned nested-JWT) expansion path.
+[MemoryDiagnoser]
+public class ActorClaimDeserializationBenchmarks
+{
+    private JsonWebTokenHandler _handler;
+    private TokenValidationParameters _validationParameters;
+    private string _tokenNoActor;
+    private string _tokenAct1;
+    private string _tokenAct2Degrade;
+    private string _tokenActort;
+
+    [GlobalSetup]
+    public void Setup()
+    {
+        _handler = new JsonWebTokenHandler();
+        JsonWebTokenHandler.MaxActorChainLength = 1;
+
+        _tokenNoActor = _handler.CreateToken(ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(0)));
+        _tokenAct1 = _handler.CreateToken(ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(1)));
+        _tokenAct2Degrade = _handler.CreateToken(ActorClaimBenchmarkData.Descriptor(ActorClaimBenchmarkData.BuildSubjectWithActorChain(2)));
+
+        // Legacy actort: an actor JWT embedded as the "actort" claim (what older JwtSecurityTokenHandler
+        // tokens look like). Read back-compat expands it into ClaimsIdentity.Actor.
+        string actorJwt = _handler.CreateToken(new SecurityTokenDescriptor
+        {
+            Subject = new CaseSensitiveClaimsIdentity(new[] { new Claim("sub", "actor-1"), new Claim("role", "worker") }),
+            SigningCredentials = BenchmarkUtils.SigningCredentialsRsaSha256,
+        });
+        _tokenActort = _handler.CreateToken(new SecurityTokenDescriptor
+        {
+            Subject = new CaseSensitiveClaimsIdentity(new[] { new Claim("sub", "user-1"), new Claim("name", "User One") }),
+            Issuer = BenchmarkUtils.Issuer,
+            Audience = BenchmarkUtils.Audience,
+            SigningCredentials = BenchmarkUtils.SigningCredentialsRsaSha256,
+            Claims = new Dictionary<string, object> { { "actort", actorJwt } },
+        });
+
+        _validationParameters = new TokenValidationParameters
+        {
+            ValidateIssuer = false,
+            ValidateAudience = false,
+            ValidateLifetime = false,
+            IssuerSigningKey = BenchmarkUtils.SigningCredentialsRsaSha256.Key,
+            ValidateIssuerSigningKey = true,
+        };
+    }
+
+    [GlobalCleanup]
+    public void Cleanup() => JsonWebTokenHandler.MaxActorChainLength = 1;
+
+    // Accessing result.ClaimsIdentity forces the (lazy) claims-identity + actor materialization.
+    [Benchmark(Baseline = true)]
+    public async Task<ClaimsIdentity> ValidateNoActor()
+    {
+        JsonWebTokenHandler.MaxActorChainLength = 1;
+        var result = await _handler.ValidateTokenAsync(_tokenNoActor, _validationParameters).ConfigureAwait(false);
+        return result.ClaimsIdentity;
+    }
+
+    [Benchmark]
+    public async Task<ClaimsIdentity> ValidateAct1()
+    {
+        JsonWebTokenHandler.MaxActorChainLength = 1;
+        var result = await _handler.ValidateTokenAsync(_tokenAct1, _validationParameters).ConfigureAwait(false);
+        return result.ClaimsIdentity;
+    }
+
+    [Benchmark]
+    public async Task<ClaimsIdentity> ValidateAct2_Degrade()
+    {
+        JsonWebTokenHandler.MaxActorChainLength = 1;
+        var result = await _handler.ValidateTokenAsync(_tokenAct2Degrade, _validationParameters).ConfigureAwait(false);
+        return result.ClaimsIdentity;
+    }
+
+    [Benchmark]
+    public async Task<ClaimsIdentity> ValidateActort_Legacy()
+    {
+        JsonWebTokenHandler.MaxActorChainLength = 1;
+        var result = await _handler.ValidateTokenAsync(_tokenActort, _validationParameters).ConfigureAwait(false);
+        return result.ClaimsIdentity;
     }
 }
